requirements.txt
apispec==5.2.2
click==8.1.3
colorama==0.4.5
Flask==2.2.2
flask-smorest==0.39.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.3
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka-python==2.0.2
MarkupSafe==2.1.1
marshmallow==3.18.0
packaging==21.3
pyparsing==3.0.9
python-dotenv==0.21.0
SQLAlchemy==1.4.41
typing_extensions==4.3.0
webargs==8.2.0
Werkzeug==2.2.2
zipp==3.8.1
